% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RLRsim-package.R
\docType{package}
\name{RLRsim-package}
\alias{RLRsim}
\alias{RLRsim-package}
\title{(Restricted) likelihood ratio tests in linear mixed models}
\description{
Rapid, simulation-based exact (restricted) likelihood ratio tests for testing
the presence of variance components/nonparametric terms with a convenient
interface for models fit with \code{nlme::lme(), lme4::lmer(), mgcv::gamm()}
or \code{SemiPar:spm()}.
}
\details{
Please see \code{\link{exactRLRT}} to get started.
}
\author{
Maintainer: Fabian Scheipl <fabian.scheipl@stat.uni-muenchen.de>

  Ben Bolker
}
\keyword{package}

