\name{RLRsim-package}
\alias{RLRsim-package}
\alias{RLRsim}
\docType{package}
\title{
(Restricted) likelihood ratio tests in linear mixed models
}
\description{Rapid, simulation-based exact (restricted) likelihood ratio tests for testing the presence of
variance components/nonparametric terms with a convenient interface for models fit with nlme::lme(),
lme4::lmer(), mgcv::gamm() and SemiPar:spm().
}
\details{
\tabular{ll}{
Package: \tab RLRsim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-01-09\cr
License: \tab GPL \cr
}
}
\author{
Maintainer: Fabian Scheipl <fabian.scheipl@stat.uni-muenchen.de>
}
\keyword{ package }
