% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RLRsim-package}
\alias{RLRsim}
\alias{RLRsim-package}
\title{(Restricted) likelihood ratio tests in linear mixed models}
\description{
Rapid, simulation-based exact (restricted) likelihood ratio tests for testing
the presence of variance components/nonparametric terms with a convenient
interface for models fit with nlme::lme(), lme4::lmer(), mgcv::gamm() and
SemiPar:spm().
}
\details{
\tabular{ll}{ Package: \tab RLRsim\cr Type: \tab Package\cr Version: \tab
2.1-1\cr Date: \tab 2013-09-17\cr License: \tab GPL \cr }
}
\author{
Maintainer: Fabian Scheipl <fabian.scheipl@stat.uni-muenchen.de>

  Ben Bolker
}
\keyword{package}

